applies changes according to github reviews.
