+++ app/src/main/java/com/orhanobut/waspsample/WaspApplication.java
@@ -40,8 +40,8 @@ public class WaspApplication extends Application {
         service = new Wasp.Builder(this)
                 .setEndpoint("http://httpbin.org")
                 .setLogLevel(LogLevel.ALL)
-                //.setHttpStack(new OkHttpStack(CertificateUtil.getTrustAllCertHttpClient()))
-                //.setHttpStack(new OkHttpStack(CertificateUtil.getPinnedCertHttpClient(this, R.raw.mytruststore, "123456")))
+                //.trustCertificates()
+                //.trustCertificates(R.raw.mytruststore, "123456")
                 .build()
                 .create(MyService.class);
